requirements.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.8
django-cors-headers==3.2.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
meson==0.52.0
pipenv==2018.11.26
psycopg2==2.7.7
pycparser==2.19
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.11.2
